Ajustes
Ajustes nas seguintes Views:
-> First Screen
-> Credits View
